requirements.txt
# =============================================================================
# Core Dependencies
# =============================================================================
# Data Processing and Analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
openpyxl>=3.1.0

# Financial Data APIs
yfinance>=0.2.36
requests>=2.31.0
# Note: Finnhub, Alpha Vantage, FRED all use REST via `requests` — no extra SDK needed

# =============================================================================
# AI / LLM
# =============================================================================
google-generativeai>=0.8.0

# =============================================================================
# Vector Databases and Memory (optional)
# =============================================================================
chromadb>=0.4.22

# =============================================================================
# Web Framework and API
# =============================================================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# =============================================================================
# Utilities
# =============================================================================
python-dotenv>=1.0.0
python-dateutil>=2.8.2
aiohttp>=3.9.0
httpx>=0.26.0

# =============================================================================
# Testing
# =============================================================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
